+++ dorae/src/main/java/com/multi/dorae/map/BookDAOInterface.java
@@ -1,15 +1,15 @@
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
 
-public interface BookServiceInterface {
+public interface BookDAOInterface {
 
 	int insert(BookVO bag);
 
 	int delete(BookVO bag);
 
 	int count(String play_id);
-	
+
 	List<BookVO> mybook(String email);
-	
+
 }
