+++ dorae/src/main/java/com/multi/dorae/map/BookService.java
@@ -1,39 +1,39 @@
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.RepositoryDefinition;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Service;
 
-@Repository
-public class BookDAO implements BookDAOInterface{
+@Service
+public class BookService implements BookServiceInterface{
+	
 	@Autowired
-	SqlSessionTemplate my;
+	BookDAOInterface dao;
 	
 	@Override
 	public int insert(BookVO bag) {
-		int result = my.insert("book.create", bag);
+		int result = dao.insert(bag);
 		return result;
 	}
 	
 	@Override
 	public int delete(BookVO bag) {
-		int result = my.delete("book.del", bag);
+		int result = dao.delete(bag);
 		return result;
 	}	
 
 	@Override
 	public int count(String play_id) {
-		int result = my.selectOne("book.cnt", play_id);
+		int result = dao.count(play_id);
 		return result;
 	}
-	
+
 	@Override
 	public List<BookVO> mybook(String email) {
-		List<BookVO> list = my.selectList("book.my", email);
+		List<BookVO> list = dao.mybook(email);
 		return list;
 	}
 }
