+++ dorae/src/main/java/com/multi/dorae/map/BookServiceInterface.java
@@ -1,15 +1,15 @@
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
 
-public interface BookDAOInterface {
+public interface BookServiceInterface {
 
 	int insert(BookVO bag);
 
 	int delete(BookVO bag);
 
 	int count(String play_id);
-
+	
 	List<BookVO> mybook(String email);
-
+	
 }
