+++ dorae/src/main/java/com/multi/dorae/map/PlayServiceInterface.java
@@ -1,4 +1,4 @@
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 public interface PlayServiceInterface {
