+++ dorae/src/main/java/com/multi/dorae/map/PreferenceDAO.java
@@ -1,11 +1,13 @@
 //PreferenceDAO.java
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+
+import com.multi.dorae.search.PlayVO;
 
 @Repository
 public class PreferenceDAO implements PreferenceDAOInterface{
@@ -21,6 +23,12 @@ public class PreferenceDAO implements PreferenceDAOInterface{
 	@Override
 	public List<PreferenceVO> mypreference(String email) {
 		List<PreferenceVO> list = my.selectList("preference.my", email);
+		return list;
+	}
+
+	@Override
+	public List<PlayVO> getRecommendedPlays(PreferenceVO bag) {
+		List<PlayVO> list = my.selectList("preference.type", bag);
 		return list;
 	}
 }
