+++ dorae/src/main/java/com/multi/dorae/map/PreferenceDAOInterface.java
@@ -1,11 +1,15 @@
 //PreferenceDAOInterface.java
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
+
+import com.multi.dorae.search.PlayVO;
 
 public interface PreferenceDAOInterface {
 	
 	int insert(PreferenceVO bag);
 	
 	List<PreferenceVO> mypreference(String email);
+	
+	List<PlayVO> getRecommendedPlays(PreferenceVO bag);
 }
