+++ dorae/src/main/java/com/multi/dorae/map/PreferenceService.java
@@ -1,10 +1,12 @@
 //PreferenceService.java
-package com.multi.dorae.book;
+package com.multi.dorae.map;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.multi.dorae.search.PlayVO;
 
 @Service
 public class PreferenceService implements PreferenceServiceInterface{
@@ -24,4 +26,10 @@ public class PreferenceService implements PreferenceServiceInterface{
 		return list;
 	}
 	
+	@Override
+	public List<PlayVO> getRecommendedPlays(PreferenceVO bag) {
+        List<PlayVO> recommendedPlays = dao.getRecommendedPlays(bag);
+        // 필요한 추가 로직을 수행하거나 필터링을 적용할 수 있습니다.
+        return recommendedPlays;
+    }
 }
